Precalculate transition points for the 10 min, 1 hour, and 10 hour marks
Store time in all quantity objects for exact offsets. While the refresh dates
are honored for user-facing refreshes, the widgets appear to be loaded 0-30s
in advance which means dynamic Date() offsets in TimeQuantity.activeRelativeDate
will be a few seconds off from the exact measurement.

By storing the date the quantity was measured (or will be measured) tightly
coupled with the value this issue is resolved.