Make sure correct metrics are shown on first login
The metric calculations and entry/category loading were done in parallel. This
meant that the metrics would be calculated prior to the full dataset existing,
so no live data would be found, and active refreshing would not occur.

This change makes the asNeeded loading blocking for initial metric calculation so
that 1/ any data that exists will be pulled if none is local 2/ local data will
be loaded before calculating the metrics.

This should not significantly slow down time to draw, as the happy state (local
cache data) will complete immediately without a network action.