Disconnect the account/catgory menus from direct control of sheets in the view
Instead of directly allowing the child components to modify bindings that will
directly show and hide sheets, use an explicit callback methods so that all final
actions are handled by the HomeView. This separation will allow some menu items
to immediately perform actions without showing a sheet.

In the previous configuration toggling states and recording events was not supported.