Listen for background store, entry, and category updates from Warehouse.
Previously the entry updates were not monitored. On open from background, some
recent entries would be refreshed and the objects would be replaced. This
disconnected a small number of entries and meant that the subscription chain
was broken.

This resolves the problem with half the UI not updating when resuming from
background, but also illustrates how fragile the connection and wrapping of
Store is by Warehouse. It's a nice separation to preserve the original interfaces,
but comes at a very high maintenance and performance cost. All data is now
resorted on every launch when the actual sorting may have (and was most likely)
already fine.

A better solution would be to split the notification for partial or complete updates
and resort in a more conservative manner. When the synced entries did not change
then no sorting is actually required. It's just a minor replacement and then a
verification of the subscriber chain.